Database-Management-Project/app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
pymongo==3.12.0
python-dotenv==0.19.0 